[VXP]: fix, download uri redirect.
