+++ VirtualApp/app/build.gradle
@@ -28,8 +28,8 @@ android {
         applicationId "io.va.exposed"
         minSdkVersion 21
         targetSdkVersion 23
-        versionCode 122
-        versionName "0.12.2"
+        versionCode 123
+        versionName "0.12.3"
         multiDexEnabled false
         android {
             defaultConfig {
