+++ VirtualApp/lib/src/main/java/com/lody/virtual/client/hook/providers/DownloadProviderHook.java
@@ -1,7 +1,10 @@
 package com.lody.virtual.client.hook.providers;
 
+import android.app.DownloadManager;
 import android.content.ContentValues;
+import android.database.Cursor;
 import android.net.Uri;
+import android.os.Bundle;
 
 import com.lody.virtual.client.core.VirtualCore;
 import com.lody.virtual.client.hook.base.MethodBox;
@@ -22,7 +25,6 @@ class DownloadProviderHook extends ExternalProviderHook {
     private static final String COLUMN_COOKIE_DATA = "cookiedata";
     private static final String COLUMN_NOTIFICATION_CLASS = "notificationclass";
     private static final String INSERT_KEY_PREFIX = "http_header_";
-
 
     private static final String[] ENFORCE_REMOVE_COLUMNS = {
             COLUMN_OTHER_UID,
@@ -58,5 +60,11 @@ class DownloadProviderHook extends ExternalProviderHook {
             }
         }
         return super.insert(methodBox, url, initialValues);
+    }
+
+    @Override
+    public Cursor query(MethodBox methodBox, Uri url, String[] projection, String selection, String[] selectionArgs, String sortOrder, Bundle originQueryArgs) throws InvocationTargetException {
+        Cursor cursor = super.query(methodBox, url, projection, selection, selectionArgs, sortOrder, originQueryArgs);
+        return new QueryRedirectCursor(cursor, DownloadManager.COLUMN_LOCAL_FILENAME);
     }
 }
