+++ VirtualApp/lib/src/main/java/com/lody/virtual/client/hook/providers/MediaProviderHook.java
@@ -1,8 +1,6 @@
 package com.lody.virtual.client.hook.providers;
 
 import android.content.ContentValues;
-import android.database.CharArrayBuffer;
-import android.database.CrossProcessCursorWrapper;
 import android.database.Cursor;
 import android.net.Uri;
 import android.os.Bundle;
@@ -22,50 +20,6 @@ class MediaProviderHook extends ProviderHook {
 
     MediaProviderHook(Object base) {
         super(base);
-    }
-
-    static class MediaCursorProxy extends CrossProcessCursorWrapper {
-
-        private int dataIndex;
-
-        /**
-         * Creates a cross process cursor wrapper.
-         *
-         * @param cursor The underlying cursor to wrap.
-         */
-        public MediaCursorProxy(Cursor cursor) {
-            super(cursor);
-            dataIndex = cursor.getColumnIndex(COLUMN_NAME);
-        }
-
-        @Override
-        public void copyStringToBuffer(int columnIndex, CharArrayBuffer buffer) {
-            if (columnIndex < 0 || columnIndex != this.dataIndex || buffer == null) {
-                super.copyStringToBuffer(columnIndex, buffer);
-                return;
-            }
-
-            String path = getString(columnIndex);
-            if (path == null) {
-                super.copyStringToBuffer(columnIndex, buffer);
-                return;
-            }
-
-            char[] chars = path.toCharArray();
-            int v1 = Math.min(chars.length, buffer.data.length);
-            System.arraycopy(chars, 0, buffer.data, 0, v1);
-            buffer.sizeCopied = v1;
-        }
-
-        @Override
-        public String getString(int columnIndex) {
-            String originalPath = super.getString(columnIndex);
-            if (columnIndex < 0 || columnIndex != this.dataIndex) {
-                return originalPath;
-            }
-            String path = NativeEngine.getEscapePath(originalPath);
-            return path;
-        }
     }
 
     @Override
@@ -92,6 +46,6 @@ class MediaProviderHook extends ProviderHook {
     @Override
     public Cursor query(MethodBox methodBox, Uri url, String[] projection, String selection, String[] selectionArgs, String sortOrder, Bundle originQueryArgs) throws InvocationTargetException {
         Cursor cursor = super.query(methodBox, url, projection, selection, selectionArgs, sortOrder, originQueryArgs);
-        return new MediaCursorProxy(cursor);
+        return new QueryRedirectCursor(cursor, COLUMN_NAME);
     }
 }
